requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
djangorestframework==3.14.0
idna==3.4
Pillow==9.2.0

python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
sendotp==0.0.4
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.12
